+++ owlapy/owl_data_ranges.py
@@ -1,12 +1,12 @@
-"""OWL data ranges
+"""OWL Data Ranges
 
 https://www.w3.org/TR/owl2-syntax/#Data_Ranges
 
 DataRange := Datatype | DataIntersectionOf | DataUnionOf | DataComplementOf | DataOneOf | DatatypeRestriction
 """
 
-from ..owl_object import OWLObject
-from ..meta_classes import HasOperands
+from .owl_object import OWLObject
+from .meta_classes import HasOperands
 from typing import Final, Sequence, Iterable
 
 from abc import ABCMeta
